project/requirements.txt
aerich==0.6.3
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.26.0
attrs==22.1.0
bcrypt==4.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
dictdiffer==0.9.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
Faker==14.2.0
fastapi==0.81.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
mccabe==0.7.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.1
pyflakes==2.5.0
pyparsing==3.0.9
pypika-tortoise==0.1.6
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tomlkit==0.11.4
tortoise-orm==0.19.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.16.0
